clearing up some messes

+++ src/main/java/com/hejazi/boostan/database/mysql/availableCoursesRepository.java
@@ -2,5 +2,5 @@ package com.hejazi.boostan.database.mysql;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface availableCoursesRepository extends CrudRepository<availableCourses, Long > {
+public interface availableCoursesRepository extends CrudRepository<availableCourses, Integer > {
 }
