+++ src/main/java/com/hejazi/boostan/database/mysql/courseRepository.java
@@ -2,5 +2,5 @@ package com.hejazi.boostan.database.mysql;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface courseRepository extends CrudRepository<course , Long >{
+public interface courseRepository extends CrudRepository<course , Integer >{
 }
