+++ src/main/java/com/hejazi/boostan/database/mysql/studentRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.Optional;
 
-public interface studentRepository extends CrudRepository<Student , Long> {
+public interface studentRepository extends CrudRepository<Student , Integer> {
     Optional<Student> findByUsername(String username);
     Optional<Student> findByPassword(String password);
 }
